Set the baseurl to empty string by default
**Why**: When Federalist is building the site, it sets the baseurl value
to nil if the site is being build with a custom domain, meaning it is
being built to be deployed to either preview.login.gov or login.gov.
This creates a problem with the multiple languages plugin which expects
the baseurl to be present. This commit adds a plugin with a hook to set
the baseurl to an empty string if it is nil, which fixes the bug.