Replace support number with contact form
We are providing a number in the filing based on this note from AWS:
~~~
Support phone number
Carriers require a support phone number for all short codes. This
number must be Local to the country where you are requesting the
short code.
~~~

Since the call in number only has a prerecorded message now we will
try to leave it out of the copy.  It might have to go back in if
we get push back from AWS or an audit.  However, it will prevent
users from wasting time calling so it is worth a try.