Add "Call us" feature to the "Contact us" page
This is the first edition addition of the "Call us" feature. It can be toggled on and off by flipping the contact_form_phone_number_enabled feature flag in _config.yml.

Otherwise, the only other minor change is that I pulled out operating hours into another field in settings, since I believe that field will be the same for both emailing and calling.